// AutoVideoMetaLocalize/client/generated-sources/openapi/models/VideoListResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AutoVideoMetaLocalize
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    PageInfo,
    PageInfoFromJSON,
    PageInfoFromJSONTyped,
    PageInfoToJSON,
    TokenPagination,
    TokenPaginationFromJSON,
    TokenPaginationFromJSONTyped,
    TokenPaginationToJSON,
    Video,
    VideoFromJSON,
    VideoFromJSONTyped,
    VideoToJSON,
} from './';

/**
 * 
 * @export
 * @interface VideoListResponse
 */
export interface VideoListResponse {
    /**
     * 
     * @type {string}
     * @memberof VideoListResponse
     */
    eTag?: string | null;
    /**
     * 
     * @type {string}
     * @memberof VideoListResponse
     */
    eventId?: string | null;
    /**
     * 
     * @type {Array<Video>}
     * @memberof VideoListResponse
     */
    items?: Array<Video> | null;
    /**
     * 
     * @type {string}
     * @memberof VideoListResponse
     */
    kind?: string | null;
    /**
     * 
     * @type {string}
     * @memberof VideoListResponse
     */
    nextPageToken?: string | null;
    /**
     * 
     * @type {PageInfo}
     * @memberof VideoListResponse
     */
    pageInfo?: PageInfo;
    /**
     * 
     * @type {string}
     * @memberof VideoListResponse
     */
    prevPageToken?: string | null;
    /**
     * 
     * @type {TokenPagination}
     * @memberof VideoListResponse
     */
    tokenPagination?: TokenPagination;
    /**
     * 
     * @type {string}
     * @memberof VideoListResponse
     */
    visitorId?: string | null;
}

export function VideoListResponseFromJSON(json: any): VideoListResponse {
    return VideoListResponseFromJSONTyped(json, false);
}

export function VideoListResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): VideoListResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'eTag': !exists(json, 'eTag') ? undefined : json['eTag'],
        'eventId': !exists(json, 'eventId') ? undefined : json['eventId'],
        'items': !exists(json, 'items') ? undefined : (json['items'] === null ? null : (json['items'] as Array<any>).map(VideoFromJSON)),
        'kind': !exists(json, 'kind') ? undefined : json['kind'],
        'nextPageToken': !exists(json, 'nextPageToken') ? undefined : json['nextPageToken'],
        'pageInfo': !exists(json, 'pageInfo') ? undefined : PageInfoFromJSON(json['pageInfo']),
        'prevPageToken': !exists(json, 'prevPageToken') ? undefined : json['prevPageToken'],
        'tokenPagination': !exists(json, 'tokenPagination') ? undefined : TokenPaginationFromJSON(json['tokenPagination']),
        'visitorId': !exists(json, 'visitorId') ? undefined : json['visitorId'],
    };
}

export function VideoListResponseToJSON(value?: VideoListResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'eTag': value.eTag,
        'eventId': value.eventId,
        'items': value.items === undefined ? undefined : (value.items === null ? null : (value.items as Array<any>).map(VideoToJSON)),
        'kind': value.kind,
        'nextPageToken': value.nextPageToken,
        'pageInfo': PageInfoToJSON(value.pageInfo),
        'prevPageToken': value.prevPageToken,
        'tokenPagination': TokenPaginationToJSON(value.tokenPagination),
        'visitorId': value.visitorId,
    };
}